src/receipt/receipt.service.ts
import { Injectable, Logger } from '@nestjs/common';

@Injectable()
export class ReceiptService {
  private readonly logger = new Logger(ReceiptService.name);
  generateReceipt(receiptDetails: any): {
    success: boolean;
    message: string;
    data;
  } {
    this.logger.log(ReceiptService.name);
    const generatedReceipt = {};
    try {
      // generate the receipt data for frontend
      this.logger.log('Receipt generated');
    } catch (error) {
      //log the error for debugging purpose
      this.logger.error(
        `Error generating receipt: ${error.message}`,
        error.stack,
        'ReceiptService',
      );

      return {
        success: false,
        message: 'Unexpected Failure',
        data: null,
      };
    }

    return {
      success: true,
      message: 'Receipt generated successfully',
      data: generatedReceipt,
    };
  }
}
